This game is actually Tic Tac Toe
In this game there is only two players who can play this game. Second one thing is that it will also shows the Turn of one player to next Player Like: If there is a turn of Player X if player X is Played then there would be turn of player O who will play the next turn.